requirements.txt
pypdf==4.0.1
requests==2.31.0
python-dotenv==1.0.0
Pillow==10.0.0
PyMuPDF==1.23.26
llama-parse==0.4.4
pdfplumber==0.11.0
google-generativeai==0.3.2 